Add files via upload
This is a REST API that I have created while doing my internship from the Rely Healthcare Technologies. This consists of the future predictions that I have obtained by performing machine learning coding on a real time dataset that consists of the details of the transactions of the pharmaceutical products from the manufacturers to the healthcare organizations. The concept I have used for obtaining the future predictions is Time Series Forecasting which is a machine learning model.